Add files via upload
The silhouette score has been improved when we perform PCA on the data set. when we 
applied kmeans on the data set without performing PCA we got a silhouette score of 46.5%. 
After performing PCA we got a silhouette score of 57%. The silhouette score has been improved 
by more than 10%. 
PCA is an unsupervised while LDA is a supervised dimensionality reduction technique.